requirements.txt
fastapi
uvicorn
langgraph
langchain
Pillow
pymupdf
pytesseract
python-multipart
pydantic
python-dotenv
pytest
